Added a simpler way to use Model to show a joined query. For that I added support for CUSTOM_SELECT_QUERY in Model: you can now indicate the type of select query you want to use for the class. This query can have two fields that are replaced in runtime:{tableName} and {tableName.fields} (which is the list of fields preceded by the tablename)
Added support for a NOT-USED: QUERY in order for certain fields not to be named directly in the previously mention {tablename.fields}, as they are filled in by the join query. Maybe not very specific, but alas.

for these fields, they also must be NOT-USED: QUERY, INSERT, UPDATE, that is their values are not used at all for inserts and updates!

changed the ModlaityClass to use a custom select query with a left join