Added PeriodicService class, generated its table and CRUD pages.
Fixed a bug where as a joined field was being generated in the table. Added CREATE to sql exceptions in order for defined fields in classes to NOT be used in the create table query.

Added an option to generfate a text field for fields marqued as DOUBLE in the class.

Added ClassAccessTemplate class, generated the table and the CRUD php files.

Added a from with a button in the Operations file to generate tables.