Adding a new RelationModel for classes that depict children or relations...probably children then. Need to refactor this.
Added a few functions for ReflectionUtil and DatabaseUtil to get which field is used for the parent id and which field is used for the child id.

Added a function to delete all children of a given parent.
Added   a function to GET all children of a given parent.

Added two tags to be used in the Doc Comment of fields : RELATION: PARENT_ID and RELATION: CHILD_ID.

The children in the cases I am using are totally contained in the parent, they have no auto increment fields, and they will be totally deleted and totally inserted whenever something changes.
this will need transactions from PDO probably or a specific function.